* add category configuration for separate output file ** update default config
todo is updated when the jira_validator runs and detects the categories

separate files will only be created for mapped values. unmapped categories go to the default bucket and the default todos.xml.

* add category configuration for separate output file
** update default config

todo is updated when the jira_validator runs and detects the categories

separate files will only be created for mapped values. unmapped categories go to the default bucket and the default todos.xml.